Added ex002 in python
